api: Add persistence for offline and page load
This adds a persistence wrapper for the read operation from the API. It
caches the data using localforage, and on startup, it cleans out
anything older than 2 days in order to keep browser storage from
growing too much. This means that the last 2 days of data read from the
application is available offline, and additionally for speeding up page
loads.